requirement.txt
Package    Version
---------- -------
pillow     10.3.0
pip        22.0.2
setuptools 59.6.0
